Add files via upload
A temporary file that contains questions 2 and 3 with Python 3.9(So there are many modifications to the written code). Orthogonal sampling is still an approximation. Variance and standard error of the mean may need to be added. Feel free to check and edit this file and I will also try to make this code fit the original one.